DNS: package refactoring and documentation
