+++ DnsService/src/main/scala/org/gammf/proxima/dns/dnsserver/package.scala
@@ -1,4 +1,4 @@
-package org.gammf.proxima.dns.communication
+package org.gammf.proxima.dns
 
 import play.api.libs.json._
 import play.api.libs.functional.syntax._
